+++ UIKit/SpecHelper/Extensions/UIControl+Spec.m
@@ -1,13 +1,21 @@
 #import "UIControl+Spec.h"
 
+static NSString *exceptionName = @"Untappable";
+static NSString *hiddenExceptionReason = @"Can't tap an invisible control";
+static NSString *disabledExceptionReason = @"Can't tap a disabled control";
+static NSString *zeroSizeExceptionReason = @"Can't tap a control with no width or height. Your control may not be laid out correctly.";
+
 @implementation UIControl (Spec)
 
 - (void)tap {
     if (self.hidden) {
-        [[NSException exceptionWithName:@"Untappable" reason:@"Can't tap an invisible control" userInfo:nil] raise];
+        [[NSException exceptionWithName:exceptionName reason:hiddenExceptionReason userInfo:nil] raise];
     }
     if (!self.isEnabled) {
-        [[NSException exceptionWithName:@"Untappable" reason:@"Can't tap a disabled control" userInfo:nil] raise];
+        [[NSException exceptionWithName:exceptionName reason:disabledExceptionReason userInfo:nil] raise];
+    }
+    if (self.bounds.size.width == 0 || self.bounds.size.height == 0) {
+        [[NSException exceptionWithName:exceptionName reason:zeroSizeExceptionReason userInfo:nil] raise];
     }
     [self sendActionsForControlEvents:UIControlEventTouchUpInside];
 }